requirements.txt
appnope==0.1.4
asttokens==3.0.0
backcall==0.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
executing==2.1.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nbimporter==0.3.4
nest-asyncio==1.6.0
numpy==1.24.4
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
rsa==4.9
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
zipp==3.20.2
